Fix deadlock in PooledClientProvider
When more threads then the size of the client pool try to get a client in parallel this either results in a deadlock (just try to run the provided test without the fixes in the PooledClientProvider) or the ElasticPool's internal queue overflows when one client too much is returned to the pool (with only the fix in the AbstractClientProvider).

Cause was that the pool was initialized twice. Once from within the pools getClientPool() method and once when creating the first client in the newClient() method.